requirements.txt
biopython==1.79
cycler==0.10.0
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.4.2
numpy==1.21.1
Pillow==8.3.1
pyparsing==2.4.7
python-dateutil==2.8.2
scikit-learn==0.24.2
scipy==1.7.1
six==1.16.0
sklearn==0.0
strkernel==0.2
threadpoolctl==2.2.0
